Refs #11 Specify poster width to avoid conflicts with site-wide image styling
It's common practice to style images site-wide with a default width of 100%. Because the poster image defines
a max-height but does not specify a width, the poster image will stretch to 100% in this case. The intent of
the poster styling is to keep the poster at a maximum height, but constrained to its original ratio.