Fixes #21 Replace webcomponents.js with webcomponents-lite.js
The `webcompontents.js` polyfill, used by the Chromecast framework,
interferes significantly with jQuery. Especially jQuery's patch to
consitantly bubble events across browsers. In addition,
`webcomponents.js` polyfills much more than is needed by the framework.

The polyfill comes in a "lite" version (`webcomponents-lite.js`). This
version of the polyfill polyfills `registerElement`, but not the shadow
DOM. With this polyfill our plugin will work without interfering with
jQuery. However, the `<google-cast-button>` custom element will not
work. Since we don't use this element, we have determined that this is
the least of all evils.

This commit changes our include of `webcomponents.js` in
`preloadWebComponents` to include `webcomponents-lite.js` instead.