Refs #23 Add readyState function to Chromecast tech
Calling the Video.js player's `readyState` function while casting causes
an error with the message:

```
VIDEOJS: Video.js: readyState method not defined for Chromecast playback
technology. TypeError: this.tech_[t] is not a function
```

So, we add a `readyState` function to the Chromecast tech that returns
the best approximation that we can make about the media's "ready state".
If the media is buffering or idle (no media is loaded), we have to
assume the readyState is "HAVE_NOTHING" because even though the player
is buffering it may not have buffered any data and we have no way to
know how much data has been buffered. In all other cases, we say that
the state is "HAVE_ENOUGH_DATA".