Refs #25 Add a `playbackRate` function
While casting an HLS media item, this error is logged in the console
periodically:

```
VIDEOJS: ERROR: TypeError: _this2.tech_.playbackRate is not a function
```

This is because the
[ChromecastTech](https://github.com/silvermine/videojs-chromecast/blob/master/src/js/tech/ChromecastTech.js)
class does not implement a `playbackRate` function. So, we add a
`playbackRate` function, as well as a setter for the `playbackRate` that
does nothing because there's no way to change the `playbackRate` using
the default receiver app.