fix: correctly fire ended event
The 'ended' event should be fired when an item ends. However, this was
only firing if the `ended` function returned true. The `ended` function,
in turn, would only return true if a MediaSession object existed and had
an `idleReason` of `FINISHED`. However, the `ended` function was never
returning `true` because the MediaSession is being destroyed before
the function is run.

To fix this, I added an update listener to the MediaSession when an item
is played. When the event is fired with a false value (indicating the
MediaSession is no longer alive) a flag is stored to show the
MediaSession has ended and this value is used in the `ended` function.