chore: run demo https tunneled from localhost
To be able to actually run the demo it needs to be served behind a
server with a valid HTTPS certificate. In order to simplify this, and
make it easier to verify and develop this library, a npm script is added
to setup the demo behind a HTTPS server in a single command.